refactors Seq -> Cont
