+++ caveat/encoding/__init__.py
@@ -9,13 +9,18 @@ from caveat.encoding.discrete import DiscreteEncoder, DiscreteEncoderPadded
 from caveat.encoding.seq2score import Seq2ScoreEncoder
 from caveat.encoding.seq2seq import Seq2SeqEncoder
 from caveat.encoding.seq_weighting import act_weight_library, seq_weight_library
-from caveat.encoding.sequence import SequenceEncoder, SequenceEncoderStaggered
+from caveat.encoding.sequence import (
+    ContinuousEncoder,
+    ContinuousEncoderStaggered,
+)
 
 library = {
     "discrete": DiscreteEncoder,
     "discrete_padded": DiscreteEncoderPadded,
-    "sequence": SequenceEncoder,
-    "sequence_staggered": SequenceEncoderStaggered,
+    "continuous": ContinuousEncoder,
+    "continous_staggered": ContinuousEncoderStaggered,
+    "sequence": ContinuousEncoder,
+    "sequence_staggered": ContinuousEncoderStaggered,
     "seq2seq": Seq2SeqEncoder,
     "seq2score": Seq2ScoreEncoder,
 }
