+++ caveat/models/continuous/vae_attention.py
@@ -5,14 +5,14 @@ import torch
 import torch.nn.functional as F
 from torch import Tensor, exp, nn
 
-from caveat.models import (
-    Base,
+from caveat.models import Base
+from caveat.models.embed import (
     CustomDurationEmbeddingAddNorm,
     CustomDurationEmbeddingConcat,
 )
 
 
-class VAESeqXAtt(Base):
+class VAEContXAtt(Base):
     def __init__(self, *args, **kwargs):
         """RNN based encoder and decoder with encoder embedding layer."""
         super().__init__(*args, **kwargs)
