+++ caveat/models/continuous/vae_cnn1d.py
@@ -1,13 +1,14 @@
-from typing import Optional, Tuple, Union
+from typing import Optional, Tuple
 
 import torch
 from torch import Tensor, nn
 
-from caveat.models import Base, CustomDurationEmbeddingConcat, utils
-from caveat.models.utils import calc_output_padding_2d, conv2d_size
+from caveat.models import Base, utils
+from caveat.models.embed import CustomDurationEmbeddingConcat
+from caveat.models.utils import calc_output_padding_1d, conv1d_size
 
 
-class VAESeqCNN2D(Base):
+class VAEContCNN1D(Base):
     def __init__(self, *args, **kwargs):
         """CNN based encoder and decoder with encoder embedding layer."""
         super().__init__(*args, **kwargs)
@@ -16,22 +17,22 @@ class VAESeqCNN2D(Base):
         hidden_layers = list
         latent_dim = int
         dropout = Optional[float]
-        kernel_size = Optional[Union[tuple[int, int], int]]
-        stride = Optional[Union[tuple[int, int], int]]
-        padding = Optional[Union[tuple[int, int], int]]
+        kernel_size = Optional[int]
+        stride = Optional[int]
+        padding = Optional[int]
 
-        encoded_size = self.encodings + 1
+        encoded_size = config.get("embed_size", self.encodings + 1)
         hidden_layers = utils.build_hidden_layers(config)
         latent_dim = config["latent_dim"]
         dropout = config.get("dropout", 0)
-        kernel_size = config.get("kernel_size", 3)
+        kernel_size = config.get("kernel_size", 2)
         stride = config.get("stride", 2)
         padding = config.get("padding", 1)
 
         self.latent_dim = latent_dim
-        # length, _ = self.in_shape
 
         self.encoder = Encoder(
+            input_encoding=self.encodings,
             encoded_size=encoded_size,
             in_shape=self.in_shape,
             hidden_layers=hidden_layers,
@@ -40,11 +41,10 @@ class VAESeqCNN2D(Base):
             stride=stride,
             padding=padding,
         )
-        # TODO add drop out to CNNs???
         self.decoder = Decoder(
             encoded_size=encoded_size,
-            target_shapes=self.encoder.target_shapes,
-            hidden_layers=hidden_layers,
+            target_shapes=self.encoder.shapes,
+            dropout=dropout,
             kernel_size=kernel_size,
             stride=stride,
             padding=padding,
@@ -75,13 +75,14 @@ class VAESeqCNN2D(Base):
 class Encoder(nn.Module):
     def __init__(
         self,
+        input_encoding: int,
         encoded_size: int,
         in_shape: tuple,
         hidden_layers: list,
         dropout: float = 0.1,
-        kernel_size: Union[tuple[int, int], int] = 3,
-        stride: Union[tuple[int, int], int] = 2,
-        padding: Union[tuple[int, int], int] = 1,
+        kernel_size: int = 2,
+        stride: int = 2,
+        padding: int = 1,
     ):
         """2d Convolutions Encoder.
 
@@ -90,51 +91,57 @@ class Encoder(nn.Module):
             in_shape (tuple[int, int, int]): [C, time_step, activity_encoding].
             hidden_layers (list, optional): _description_. Defaults to None.
             dropout (float): dropout. Defaults to 0.1.
-            kernel_size (Union[tuple[int, int], int], optional): _description_. Defaults to 3.
-            stride (Union[tuple[int, int], int], optional): _description_. Defaults to 2.
-            padding (Union[tuple[int, int], int], optional): _description_. Defaults to 1.
+            kernel_size (int): kernel size. Defaults to 2.
+            stride (int): stride. Defaults to 2.
+            padding (int): padding. Defaults to 1.
         """
         super(Encoder, self).__init__()
-        h = in_shape[0]
+        print(in_shape)
+        length = in_shape[0]
         self.embedding = CustomDurationEmbeddingConcat(
-            encoded_size, encoded_size, dropout=dropout
+            input_encoding, encoded_size, dropout=dropout
         )
-        w = encoded_size
-        channels = 1
 
+        channels = encoded_size
+        self.shapes = []
         modules = []
-        self.target_shapes = [(channels, h, w)]
 
         for hidden_channels in hidden_layers:
+            self.shapes.append((channels, length))
+            if length + padding < kernel_size:
+                print("Skipping convolution:", length, kernel_size)
+                break
             modules.append(
                 nn.Sequential(
-                    nn.Conv2d(
+                    nn.Conv1d(
                         in_channels=channels,
                         out_channels=hidden_channels,
                         kernel_size=kernel_size,
                         stride=stride,
                         padding=padding,
-                        # bias=False,
+                        bias=False,
                     ),
-                    nn.BatchNorm2d(hidden_channels),
+                    nn.BatchNorm1d(hidden_channels),
                     nn.LeakyReLU(),
+                    nn.Dropout(dropout),
                 )
             )
-            h, w = conv2d_size(
-                (h, w), kernel_size=kernel_size, padding=padding, stride=stride
+            length = conv1d_size(
+                length=length,
+                kernel_size=kernel_size,
+                stride=stride,
+                padding=padding,
             )
-            self.target_shapes.append((hidden_channels, h, w))
             channels = hidden_channels
+        self.shapes.append((channels, length))
 
-        self.dropout = nn.Dropout(dropout)
-
-        self.shape_before_flattening = (-1, channels, h, w)
         self.encoder = nn.Sequential(*modules)
-        self.flat_size = int(channels * h * w)
+        self.shape_before_flattening = (-1, channels, length)
+        self.flat_size = int(channels * length)
 
     def forward(self, x):
-        y = self.dropout(self.embedding(x.int()))
-        y = y.unsqueeze(1)  # add channel dim for Conv
+        y = self.embedding(x.int())
+        y = y.permute(0, 2, 1)
         y = self.encoder(y)
         y = y.flatten(start_dim=1)
         return y
@@ -145,58 +152,57 @@ class Decoder(nn.Module):
         self,
         encoded_size: int,
         target_shapes: list,
-        hidden_layers: list,
-        kernel_size: Union[tuple[int, int], int] = 3,
-        stride: Union[tuple[int, int], int] = 2,
-        padding: Union[tuple[int, int], int] = 1,
+        dropout: float = 0.1,
+        kernel_size: int = 3,
+        stride: int = 2,
+        padding: int = 0,
     ):
-        """2d Conv Decoder.
+        """1d Conv Decoder.
 
         Args:
-            target_shapes (list): list of target shapes from encoder.
-            hidden_layers (list, optional): _description_. Defaults to None.
-            kernel_size (Union[tuple[int, int], int], optional): _description_. Defaults to 3.
-            stride (Union[tuple[int, int], int], optional): _description_. Defaults to 2.
-            padding (Union[tuple[int, int], int], optional): _description_. Defaults to 1.
+            encoded_size (int): number of encoding classes and hidden size.
+            target_shapes (list): list of target shapes.
+            dropout (float): dropout. Defaults to 0.1.
+            kernel_size (int): kernel size. Defaults to 3.
+            stride (int): stride. Defaults to 2.
+            padding (int): padding. Defaults to 0.
         """
         super(Decoder, self).__init__()
         self.hidden_size = encoded_size
         modules = []
         target_shapes.reverse()
 
-        for i in range(len(hidden_layers) - 1):
-            modules.append(
-                nn.Sequential(
-                    nn.ConvTranspose2d(
-                        in_channels=target_shapes[i][0],
-                        out_channels=target_shapes[i + 1][0],
-                        kernel_size=kernel_size,
-                        stride=stride,
-                        padding=padding,
-                        output_padding=calc_output_padding_2d(
-                            target_shapes[i + 1]
-                        ),
-                        # bias=False,
-                    ),
-                    nn.BatchNorm2d(target_shapes[i + 1][0]),
-                    nn.LeakyReLU(),
+        for i in range(len(target_shapes) - 1):
+            c_in, l_in = target_shapes[i]
+            c_out, l_out = target_shapes[i + 1]
+            if c_in == c_out and l_in == l_out:
+                print(
+                    "Skipping transpose convolution:", c_in, l_in, c_out, l_out
                 )
+                continue
+            in_padding, out_padding = calc_output_padding_1d(
+                length=l_in,
+                target=l_out,
+                kernel_size=kernel_size,
+                stride=stride,
+                padding=padding,
             )
-
-        # Final layer with Tanh activation
-        modules.append(
-            nn.Sequential(
-                nn.ConvTranspose2d(
-                    in_channels=target_shapes[-2][0],
-                    out_channels=target_shapes[-1][0],
+            block = [
+                nn.ConvTranspose1d(
+                    in_channels=c_in,
+                    out_channels=c_out,
                     kernel_size=kernel_size,
                     stride=stride,
-                    padding=padding,
-                    output_padding=calc_output_padding_2d(target_shapes[-1]),
+                    padding=in_padding,
+                    output_padding=out_padding,
+                    bias=False,
                 ),
-                nn.BatchNorm2d(target_shapes[-1][0]),
-            )
-        )
+                nn.BatchNorm1d(c_out),
+            ]
+            if i < len(target_shapes) - 2:
+                block.append(nn.LeakyReLU())
+                block.append(nn.Dropout(dropout))
+            modules.append(nn.Sequential(*block))
 
         self.decoder = nn.Sequential(*modules)
         self.logprob_activation = nn.LogSoftmax(dim=-1)
@@ -204,7 +210,7 @@ class Decoder(nn.Module):
 
     def forward(self, hidden, **kwargs):
         y = self.decoder(hidden)
-        y = y.squeeze(1)  # remove conv channel dim
+        y = y.permute(0, 2, 1)
         acts_logits, durations = torch.split(
             y, [self.hidden_size - 1, 1], dim=-1
         )
