+++ configs/conditional_paper/c_discrim.yaml
@@ -22,27 +22,27 @@ experiment_params:
   weight_decay: 0.02
   scheduler_gamma: 0.975
 
-conditionals:
-  gender: nominal
-  age_group: nominal
-  car_access: nominal
-  work_status: nominal
-  income: nominal
-
 evaluation_params:
   split_on: [gender, age_group, car_access, work_status, income]
 
-attribute_encoder: "tokens"
+labels_encoder:
+  name: "tokens"
+  labels:
+    gender: nominal
+    age_group: nominal
+    car_access: nominal
+    work_status: nominal
+    income: nominal
 
 encoder_params:
-  name: "sequence"
+  name: "continuous"
   max_length: 16
   norm_duration: 1440
   jitter: 0.3
   fix_durations: True
   
 model_params:
-  name: "CondSeqLSTM"
+  name: "CondContLSTM"
   latent_dim: 0
   hidden_n: 4
   hidden_size: 128
