--- configs/cvaes_eval.yaml
@@ -1,55 +1,0 @@
-global:
-  logging_params:
-    log_dir: "logs/cvaes"
-    name: "batch"
-
-  schedules_path: "tmp/nts_home_schedules.csv"
-  attributes_path: "tmp/nts_home_attributes_groups.csv"
-
-  conditionals:
-    gender: nominal
-    age_group: nominal
-    car_access: nominal
-    work_status: nominal
-    income: nominal
-
-  evaluation_params:
-    # split_on: [gender, age_group, car_access, work_status, income]
-    split_on: [work_status]
-
-  attribute_encoder: "tokens"
-
-  encoder_params:
-    name: "sequence"
-    max_length: 12
-    norm_duration: 1440
-    jitter: 0.1
-
-  loader_params:
-    train_batch_size: 1024
-    val_batch_size:  1024
-    num_workers: 4
-
-# Feed:
-#   seed: 11111
-
-# FeedPre:
-#   seed: 22222
-
-# Latent:
-#   seed: 33333
-
-# LatentPre:
-#   seed: 44444
-
-# LatentFeed:
-#   seed: 12345
-
-LatentFeedPre:
-  seed: 12345
-
-# LatentAddFeed:
-#   seed: 12345
-
-LatentAddFeedPre:
-  seed: 12345
