+++ configs/examples/mmrun.yaml
@@ -10,13 +10,13 @@ labels_encoder:
     work_status: nominal
 
 encoder_params:
-  name: "sequence"
+  name: "continuous"
   max_length: 12
   norm_duration: 1440
   jitter: 0.2
 
 model_params:
-  name: "VAESeqLSTM"
+  name: "VAEContLSTM"
   hidden_n: 4
   hidden_size: 128
   latent_dim: 6
